move stdoutmute
